fix: 피드백 내용 반영 1 문자를 구분하는 SplitFactory 숫자를 검증하는 NumberValidator 전체의 합을 담당하는 StringAddCalculator 로 역할분리 Error 형식을 IllegalArgumentException 으로 구체화
